+++ src/main/java/com/inq/kubs/domain/member/dto/request/CreateMemberRequest.java
@@ -1,0 +1,19 @@
+package com.inq.kubs.domain.member.dto.request;
+
+import lombok.Data;
+
+@Data
+public class CreateMemberRequest {
+
+    private String name;
+
+    private Long studentId;
+
+    private String pw;
+
+    private String email;
+
+    private String phoneNumber;
+
+    private Long departmentId;
+}
