+++ src/main/java/com/inq/kubs/domain/member/service/MemberService.java
@@ -7,6 +7,7 @@ import com.inq.kubs.domain.member.dto.CreateMemberDto;
 import com.inq.kubs.domain.member.repository.MemberRepository;
 import com.inq.kubs.web.exception.ErrorType;
 import com.inq.kubs.web.exception.KubsException;
+import com.inq.kubs.domain.member.dto.request.CreateMemberRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,11 +21,12 @@ public class MemberService {
     private final DepartmentRepository departmentRepository;
 
     @Transactional
-    public Long createMember(CreateMemberDto dto, Long departmentId) {
+    public Long createMember(CreateMemberRequest request) {
 
-        Department department = departmentRepository.findById(departmentId)
+        Department department = departmentRepository.findById(request.getDepartmentId())
                 .orElseThrow(() -> new KubsException(ErrorType.NOT_EXIST_KEY));
 
+        CreateMemberDto dto = new CreateMemberDto(request);
         dto.registerDepartment(department);
         Member member = Member.createMember(dto);
         memberRepository.save(member);
